Rewrite install_platform.sh (include /opt/Xilinx/aws-fpga)
* Include the aws-fpga project in /opt/Xilinx so it is available at
  runtime.
* Rewrite script to put configurables at the top.
* Include many more comments about what is going on.
* Use long form command line arguments where available.
* Drop redundant sdk_install.sh call (included via sdaccel_setup.sh).

Old state:

We used to only take a temporary copy of aws-fpga, and copy out the
development kit, deleting it when done. However, AWS now requires that
sdaccel_setup.sh has been sourced in order to do AFI generation.
Therefore, put it in /opt/Xilinx, so that it takes advantage of the
existing bind mounts and is available at runtime inside our builder
docker containers.

We used to source sdk_install.sh. This seems obviously incorrect -
according to that script it is not meant to be sourced. It is also run
as a part of sdaccel_setup.sh, which sources sdk_setup.sh, and
eventually runs sdk_install.sh, so it looks unnecessary.